requirements.txt
langchain
streamlit
openai
pydantic
streamlit-authenticator
deta == 1.1.0
python-decouple
